+++ src/test/java/io/github/astrapi69/random/RandomCharactersTest.java
@@ -24,9 +24,9 @@
  */
 package io.github.astrapi69.random;
 
-import static org.testng.AssertJUnit.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import org.testng.annotations.Test;
+import org.junit.jupiter.api.Test;
 
 import io.github.astrapi69.test.base.AbstractTestCase;