+++ src/test/java/io/github/astrapi69/random/number/RandomCharFactoryTest.java
@@ -24,14 +24,14 @@
  */
 package io.github.astrapi69.random.number;
 
-import static org.testng.Assert.assertEquals;
-import static org.testng.AssertJUnit.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 import java.nio.CharBuffer;
 import java.security.SecureRandom;
 
+import org.junit.jupiter.api.Test;
 import org.meanbean.test.BeanTester;
-import org.testng.annotations.Test;
 
 import io.github.astrapi69.random.DefaultSecureRandom;
 import io.github.astrapi69.random.RandomCharacters;