+++ src/test/java/io/github/astrapi69/random/number/RandomNumberExtensionsTest.java
@@ -24,13 +24,13 @@
  */
 package io.github.astrapi69.random.number;
 
-import static org.testng.AssertJUnit.assertNotNull;
-import static org.testng.AssertJUnit.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 import org.meanbean.test.BeanTester;
-import org.testng.annotations.AfterMethod;
-import org.testng.annotations.BeforeMethod;
-import org.testng.annotations.Test;
 
 import io.github.astrapi69.string.StringExtensions;
 import io.github.astrapi69.test.base.BaseTestCase;
@@ -50,7 +50,7 @@ public class RandomNumberExtensionsTest extends BaseTestCase
 	 * {@inheritDoc}
 	 */
 	@Override
-	@BeforeMethod
+	@BeforeEach
 	protected void setUp() throws Exception
 	{
 		super.setUp();
@@ -60,7 +60,7 @@ public class RandomNumberExtensionsTest extends BaseTestCase
 	 * {@inheritDoc}
 	 */
 	@Override
-	@AfterMethod
+	@AfterEach
 	protected void tearDown() throws Exception
 	{
 		super.tearDown();
@@ -69,7 +69,7 @@ public class RandomNumberExtensionsTest extends BaseTestCase
 	/**
 	 * Test method for {@link RandomNumberExtensions#getRandomNumberString(int, int)}
 	 */
-	@Test(enabled = true)
+	@Test
 	public void testGetRandomFloatString()
 	{
 		int beforeComma;
@@ -86,7 +86,7 @@ public class RandomNumberExtensionsTest extends BaseTestCase
 	/**
 	 * Test method for {@link RandomNumberExtensions#getRandomNumericString()}
 	 */
-	@Test(enabled = true)
+	@Test
 	public void testGetRandomNumericString()
 	{
 		String randomNumericString = RandomNumberExtensions.getRandomNumericString();