+++ src/test/java/io/github/astrapi69/random/object/RandomWebObjectFactoryTest.java
@@ -24,16 +24,17 @@
  */
 package io.github.astrapi69.random.object;
 
-import static org.testng.Assert.assertEquals;
-import static org.testng.Assert.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 import java.nio.CharBuffer;
 import java.util.Optional;
 
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 import org.meanbean.test.BeanTester;
-import org.testng.annotations.AfterMethod;
-import org.testng.annotations.BeforeMethod;
-import org.testng.annotations.Test;
 
 import io.github.astrapi69.random.RandomCharacters;
 import io.github.astrapi69.test.base.BaseTestCase;
@@ -53,7 +54,7 @@ public class RandomWebObjectFactoryTest extends BaseTestCase
 	 * {@inheritDoc}
 	 */
 	@Override
-	@BeforeMethod
+	@BeforeEach
 	public void setUp() throws Exception
 	{
 	}
@@ -62,7 +63,7 @@ public class RandomWebObjectFactoryTest extends BaseTestCase
 	 * {@inheritDoc}
 	 */
 	@Override
-	@AfterMethod
+	@AfterEach
 	public void tearDown() throws Exception
 	{
 	}
@@ -109,10 +110,13 @@ public class RandomWebObjectFactoryTest extends BaseTestCase
 	/**
 	 * Test method for {@link RandomWebObjectFactory#getInfomailFromWebsite(java.lang.String)} .
 	 */
-	@Test(expectedExceptions = IllegalArgumentException.class)
+	@Test
 	public void testGetInfomailFromWebsiteExEx()
 	{
-		RandomWebObjectFactory.getInfomailFromWebsite("htp://ww.g.rw");
+		Assertions.assertThrows(IllegalArgumentException.class, () -> {
+			// ...
+			RandomWebObjectFactory.getInfomailFromWebsite("htp://ww.g.rw");
+		});
 	}
 
 	/**